pyproject.toml
[tool.poetry]
name = "pdf-bot"
version = "1.0.0"
description = "Telegram bot that can do a lot of things related to PDF files."
authors = ["zeshuaro <zeshuaro@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
Babel = "2.11.0"
google-cloud-datastore = "2.11.0"
humanize = "4.4.0"
img2pdf = "0.4.4"
ocrmypdf = "14.0.1"
pdf2image = "==1.16.1"
"pdfminer.six" = "20221105"
python-dotenv = "0.21.0"
python-telegram-bot = { extras = [
    "callback-data",
    "rate-limiter",
    "webhooks",
], version = "20.0b0" }
Pillow = "9.3.0"
PyPDF2 = "2.12.1"
requests = "2.28.1"
slack-sdk = "3.19.5"
weasyprint = "57.2"
noteshrink = { git = "https://github.com/zeshuaro/noteshrink" }
pdf-diff = { git = "https://github.com/zeshuaro/pdf-diff" }
langdetect = "1.0.9"
loguru = "0.6.0"
sentry-sdk = "1.12.1"
dependency-injector = "4.41.0"
pdfCropMargins = "==1.1.9"
pycryptodome = "3.16.0"
pydantic = { extras = ["dotenv"], version = "==1.10.4" }

[tool.poetry.group.dev.dependencies]
black = "22.12.0"
isort = "5.11.4"
pylint = "2.15.9"
pytest = "7.2.0"
pytest-cov = "4.0.0"
pre-commit = "==2.21.0"
mypy = "0.991"
types-requests = "2.28.11.7"
pytest-asyncio = "0.20.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pylint.master]
extension-pkg-allow-list = ["dependency_injector", 'pydantic']

[tool.pylint.message-control]
disable = [
    "consider-using-f-string",
    "consider-using-with",
    "duplicate-code",
    "invalid-name",
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
    "too-few-public-methods",
    "too-many-arguments",
    "too-many-branches",
    "too-many-instance-attributes",
    "too-many-locals",
    "too-many-return-statements",
    "unspecified-encoding",
]

[tool.pylint.tests]
disable = [
    "attribute-defined-outside-init",
    "too-many-public-methods",
    "protected-access",
]

[tool.mypy]
warn_unused_configs = true
warn_redundant_casts = true
show_error_codes = true

disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true

warn_unused_ignores = true
warn_no_return = true
warn_return_any = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ['pdf_bot.log.*']
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    'weasyprint.*',
    'langdetect.*',
    'img2pdf.*',
    'pdf2image.*',
    'pdf_diff.*',
    'noteshrink.*',
    'google.*',
]
ignore_missing_imports = true

[tool.coverage.run]
relative_files = true
branch = true
source = ["pdf_bot"]
omit = ["tests/*", 'pdf_bot/log/*']

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]
